+++ replacer-backend/replacer-domain/src/main/java/es/bvalero/replacer/finder/Replacement.java
@@ -23,6 +23,9 @@ import org.springframework.lang.NonNull;
 @Builder
 public class Replacement implements FinderResult {
 
+    public static final int CONTEXT_THRESHOLD = 20;
+    private static final int MAX_CONTEXT_LENGTH = 255; // Constrained by the database
+
     @NonNull
     FinderPage page;
 
@@ -76,6 +79,9 @@ public class Replacement implements FinderResult {
     }
 
     public String getContext() {
-        return ReplacerUtils.getContextAroundWord(page.getContent(), start, getEnd(), 20);
+        return StringUtils.truncate(
+            ReplacerUtils.getContextAroundWord(page.getContent(), start, getEnd(), CONTEXT_THRESHOLD),
+            MAX_CONTEXT_LENGTH
+        );
     }
 }