+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Remove feature: Public IP
 - Make Admin menu in frontend a dropdown
 - Fix: reload empty cache to find pages left to review after
+- Consider replacements with the same context but far enough as different
 
 ## 2.23.6 (8 May 2023)
 - Fix regression: dropdown to change language